+++ router.d.ts
@@ -10,6 +10,39 @@ export type Route = {
     segments: Array<string>;
 };
 
+export type RedirectingDetails = {
+    path: string,
+    hash: string,
+    params: Params,
+};
+export type PreloadingDetails = {
+    path: string,
+    hash: string,
+    params: Params,
+};
+export type LoadingDetails = {
+    path: string,
+    hash: string,
+    params: Params,
+    tokens: Tokens,
+    data: Data,
+};
+
+export type LoadedDetails = {
+    path: string,
+    hash: string,
+    tokens: Tokens,
+    params: Params,
+    data: Data,
+};
+
+export type OutgoingDetails = {
+    path: string,
+    hash: string,
+    params: Params,
+    tokens: Tokens,
+};
+
 export type Module = {
     tagName: string;
     file: string;
@@ -21,6 +54,10 @@ export type Tokens = {
 
 export type Params = {
     [param: string]: string | Array<string>;
+};
+
+export type Data = {
+    [key:string]: any;
 };
 
 export type GroupSettings = {
@@ -60,6 +97,9 @@ declare class Router {
         url: string,
         middleware?: Array<Function>
     ): void;
+    public enableTransitions(): void;
+    public disableTransitions(): void;
+    public setTransitionTimer(ms:number|null): void;
 }
 
 declare const mount: (element: HTMLElement) => void;
@@ -68,3 +108,4 @@ declare const pageJump: (hash: string, jump?: "auto" | "smooth") => void;
 declare const replaceState: (url: string) => void;
 declare const pushState: (url: string) => void;
 declare const router: Router;
+declare const transition: () => void;